Python-Service/requirements.txt
Flask
pillow
pytesseract
opencv-python
numpy
google-generativeai
Werkzeug
easyocr
pandas